requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
black==23.12.1
boto3==1.37.14
botocore==1.37.14
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
coverage==7.6.12
cron-descriptor==1.4.5
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.20
django-allauth==0.63.6
django-anymail==12.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-dotenv==1.4.2
django-environ==0.12.0
dj-database-url==2.1.0
environs==11.0.0
django-filter==23.5
django-import-export==4.3.7
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-safedelete==1.4.1
django-storages==1.6.5
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.9
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==37.0.0
flake8==6.1.0
getkey==0.6.5
gunicorn==21.2.0
humanize==4.12.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
phonenumbers==9.0.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.21.0
pyflakes==3.1.0
pygame==2.6.1
PyJWT==2.9.0
pyotp==2.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pytest==7.4.4
pytest-cov==6.0.0
pytest-django==4.10.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
s3transfer==0.11.4
sentry-sdk==1.45.1
setuptools==75.8.2
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
xlrd==2.0.1
xlwt==1.3.0
